+++ src/rpclib/const/http.py
@@ -1,0 +1,26 @@
+
+#
+# rpclib - Copyright (C) Rpclib contributors.
+#
+# This library is free software; you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public
+# License as published by the Free Software Foundation; either
+# version 2.1 of the License, or (at your option) any later version.
+#
+# This library is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public
+# License along with this library; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+#
+
+HTTP_200 = '200 OK'
+HTTP_400 = '400 Bad Request'
+HTTP_404 = '404 Method Not Found'
+HTTP_405 = '405 Method Not Allowed'
+HTTP_413 = '413 Request Too Long'
+HTTP_500 = '500 Internal Server Error'
+HTTP_501 = '501 Not Implemented'